# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: بذور Seeds Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: onSale
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders (admin gets all, user gets own)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithPayment"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, customer]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
...
    Category:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
        productCount:
          type: integer
      required:
        - id
        - nameAr
        - nameEn
        - slug
...
    Product:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        price:
          type: number
        salePrice:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        quantity:
          type: integer
        sku:
          type: string
        featured:
          type: boolean
        onSale:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - price
        - quantity
    CreateProductRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        price:
          type: number
        salePrice:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        quantity:
          type: integer
        sku:
          type: string
        featured:
          type: boolean
        onSale:
          type: boolean
      required:
        - nameAr
        - nameEn
        - price
        - quantity
    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productNameAr:
          type: string
        productNameEn:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
        transactionId:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        shippingZoneId:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - items
        - total
        - status
        - paymentStatus
    OrderWithPayment:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        clientSecret:
          type: string
      required:
        - order
    ShippingAddress:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        city:
          type: string
        area:
          type: string
        street:
          type: string
        building:
          type: string
      required:
        - name
        - phone
        - city
    CreateOrderRequest:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        shippingZoneId:
          type: string
        paymentMethod:
          type: string
          enum: [stripe, cash_on_delivery]
        offerCode:
          type: string
        notes:
          type: string
      required:
        - shippingAddress
        - paymentMethod
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      required:
        - status
    OrdersListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
...