# lib/api-zod/src/generated/types/ordersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * بذور Seeds Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export interface OrdersListResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}